requirements.txt
django
djangorestframework
selenium
chromedriver-binary==73.0.3683.68